Work in progress: user authentication using filters.
Work in progress: continuation

Work in progress: database

WIP: association tables created.

WIP: continuation

Database WIP almost done.

Work in progress:bug correction

Entity manager in progress;

Entity classes and their managers finished. Current existing details in managers are deemed to be enough for testing purposes; the architecture is light in such a way that it's easy to extend them.

hibernate configuration file no longer ignored (see added entity mapping). To use it for local test you have to modify db credentials so that they correspond to the one you have locally; with just one preference to the naming of the database (name it treasure).

Resources files removed with the view of replacing them with the ones in the master branch after rebasing.